report v2.0 & reservation v3.1 & index v3.2 + errPages

+++ PlaninarskoDrustvo/src/main/java/com/pd/controller/UlogeController.java
@@ -1,5 +1,6 @@
 package com.pd.controller;
 
+import java.io.IOException;
 import java.security.Principal;
 import java.text.DateFormat;
 import java.text.ParseException;
@@ -19,14 +20,18 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.pd.repository.ClanarinaRepository;
 import com.pd.repository.DomRepository;
+import com.pd.repository.IzvestajRepository;
 import com.pd.repository.KomentariseRepository;
 import com.pd.repository.KorisnikRepository;
 import com.pd.repository.ObilaziRepository;
 import com.pd.repository.PlaninaRepository;
 import com.pd.repository.RezerviseRepository;
+import com.pd.repository.SlikaRepository;
 import com.pd.repository.TerminRepository;
 import com.pd.repository.UlogaRepository;
 import com.pd.repository.ZakazujeRepository;
@@ -35,11 +40,13 @@ import com.pd.security.GeneratePassword;
 
 import model.Clanarina;
 import model.Dom;
+import model.Izvestaj;
 import model.Komentarise;
 import model.Korisnik;
 import model.Obilazi;
 import model.Planina;
 import model.Rezervise;
+import model.Slika;
 import model.Termin;
 import model.Uloga;
 import model.Zakazuje;
@@ -75,6 +82,10 @@ public class UlogeController {
 	TerminRepository trepo;
 	@Autowired
 	ZakazujeRepository zakrepo;
+	@Autowired
+	IzvestajRepository irepo;
+	@Autowired
+	SlikaRepository srepo;
 	
 	
 	
@@ -89,7 +100,7 @@ public class UlogeController {
 		
 		//TODO: Proveri da li je korisniku, ako je Clan, istekla clanarina... ako jest izbrisi mi Clanarinu i ULOGU
 		
-		if(k.getUloga().getTip().equals("Sekretar")) {
+		if(k.getUloga() != null && k.getUloga().getTip().equals("Sekretar")) {
 			List<Korisnik> korisnici = krepo.findAll();
 			List<Korisnik> zahtevi = korisnici.stream()
 					.filter(tmp -> tmp.getUloga() == null)
@@ -99,9 +110,12 @@ public class UlogeController {
 					.filter(tmp -> tmp.getUloga().getTip().equals("Planinar"))
 					.collect(Collectors.toList());
 			List<Planina> planine = prepo.findAll();
+			List<Znamenitost> sveZnamenitosti = zrepo.findAll();
+			sveZnamenitosti = sveZnamenitosti.stream().filter(z -> z.getZakazujeSe()).collect(Collectors.toList());;
 			request.getSession().setAttribute("zahtevi", zahtevi);
 			request.getSession().setAttribute("clanovi", clanovi);
 			request.getSession().setAttribute("planine", planine);
+			request.getSession().setAttribute("sveZnamenitosti", sveZnamenitosti);
 		}
 		
 		return "index";
@@ -122,7 +136,73 @@ public class UlogeController {
 	
 	@GetMapping("user/getReport")
 	public String getReport() {
+		List<Planina> planine = prepo.findAll();
+		request.getSession().setAttribute("planine", planine);
+		
+		
 		return "report";
+	}
+	
+	@GetMapping("/user/izvestajiZaPlaninu")
+	public String getIzvestajiZaPlaninu(String idP) {
+		Planina p = prepo.findById(Integer.parseInt(idP)).get();
+		
+		List<Rezervise> tmp = p.getDoms().stream().map(d -> d.getRezervises()).flatMap(List::stream).collect(Collectors.toList());
+		List<Izvestaj> izvestaji = tmp.stream().map(r -> r.getIzvestajs()).flatMap(List::stream).collect(Collectors.toList());
+		
+		request.getSession().setAttribute("izvestaji", izvestaji);
+		request.getSession().setAttribute("trazenaPlanina", p);
+		
+		return "report";
+	}
+	
+	@PostMapping("/user/postaviIzvestaj") 
+	public String postaviIzvesta(String izvestaj, String idP, String username) {
+		Korisnik k = krepo.findByKorisnickoIme(username).get();
+		Planina p = prepo.findById(Integer.parseInt(idP)).get();
+		
+		p.getDoms().forEach(d -> {
+			d.getRezervises().forEach(r -> {
+				if(r.getKorisnik().equals(k)) {
+					Izvestaj izv = new Izvestaj();
+					izv.setRezervise(r);
+					izv.setSadrzaj(izvestaj);
+					irepo.save(izv);
+					r.getIzvestajs().add(izv);
+				}
+			});
+		});
+		
+		return "redirect:/user/izvestajiZaPlaninu?idP="+idP;
+	}
+	
+	@PostMapping("/user/postaviSliku") 
+	public String postaviSliku(@RequestParam("file") MultipartFile file, String idP, String username) throws IOException {
+		
+		System.err.println(file);
+		byte[] b = file.getBytes();
+		
+		Korisnik k = krepo.findByKorisnickoIme(username).get();
+		Planina p = prepo.findById(Integer.parseInt(idP)).get();
+		
+		p.getDoms().forEach(d -> {
+			d.getRezervises().forEach(r -> {
+				if(r.getKorisnik().equals(k)) {
+					Izvestaj izv = new Izvestaj();
+					
+					
+					izv.setRezervise(r);
+					irepo.save(izv);
+
+					Slika sl = new Slika();
+					sl.setIzvestaj(izv);
+					sl.setSlika(b);
+					srepo.save(sl);
+				}
+			});
+		});
+		
+		return "redirect:/user/izvestajiZaPlaninu?idP="+idP;
 	}
 	
 	@GetMapping("/user/getSights")
@@ -138,7 +218,7 @@ public class UlogeController {
 	}
 	
 	@GetMapping("/user/pretragaPlanana")
-	public String pretragaPlanina(String idPlanina, String date) throws ParseException {
+	public String pretragaPlanina(String idPlanina, String date, String brDana) throws ParseException {
 		DateFormat df = new SimpleDateFormat("yyyy-MM-dd");
 		Date result =  df.parse(date);  
 		int id = Integer.parseInt(idPlanina);
@@ -147,15 +227,14 @@ public class UlogeController {
 		request.getSession().setAttribute("planina", p);
 		
 		request.getSession().setAttribute("datumRezervacije", result);
-		//TODO: Dobaviti informaciju koliko korisnik ostaje...
+		request.getSession().setAttribute("brDana", brDana);
 		
 		return "reservation";
 	}
 	
 	@GetMapping("/user/rezervisiDom")
 	public String rezervisiDom(String idD, String username) {
-		//TODO: Dobaviti broj dana koliko ostaje...
-		int brojDana = 7;
+		int brojDana = Integer.parseInt((String) request.getSession().getAttribute("brDana"));
 		Korisnik k = krepo.findByKorisnickoIme(username).get();
 		Dom d = drepo.findById(Integer.parseInt(idD)).get();
 		
@@ -196,32 +275,37 @@ public class UlogeController {
 	
 //  Treba da se prosledjuje USERNAME (javlja sa u indexu i u znamenitosti...)
 	@GetMapping("/user/posecuje")
-	public String posetiZnamenitost(String idZ, String idK) {
+	public String posetiZnamenitost(String idZ, String idK) throws Exception {
 		Integer idZnamenitost = Integer.parseInt(idZ);
 		Integer idKorisnik = Integer.parseInt(idK);
 		
 		Znamenitost z = zrepo.findById(idZnamenitost).get();
 		Korisnik k = krepo.findById(idKorisnik).get();
+		boolean ok = true;
 
 		Obilazi o = new Obilazi();
-		z.getStaza().getPlanina().getDoms().forEach(d -> {
-			d.getRezervises().forEach(r -> {
+		for (Dom d : z.getStaza().getPlanina().getDoms()) {
+			for (Rezervise r : d.getRezervises()) {
 				if(r.getKorisnik().equals(k)) {
-					o.setRezervise(r);
-					r.getObilazis().add(o);
-					rrepo.save(r);
-					if(z.getZakazujeSe()) {
-						z.getTermins().forEach(t -> {
-							t.getZakazujes().forEach(zak -> {
-								if(zak.getRezervise().equals(r)) {
-									zakrepo.delete(zak);
-								}
+					if(r.getPocetak().before(new Date())) {
+						o.setRezervise(r);
+						r.getObilazis().add(o);
+						rrepo.save(r);
+						if(z.getZakazujeSe()) {
+							z.getTermins().forEach(t -> {
+								t.getZakazujes().forEach(zak -> {
+									if(zak.getRezervise().equals(r)) {
+										zakrepo.delete(zak);
+									}
+								});
 							});
-						});
+						}
+					} else {
+						return "tmp/errorPosecuje";
 					}
 				}
-			});
-		});
+			}
+		}
 		o.setZnamenitost(z);
 		orepo.save(o);
 		return "redirect:/guest";
@@ -285,6 +369,12 @@ public class UlogeController {
 		return "redirect:/guest";
 	}
 	
+	@GetMapping("/admin/generisiIzvestaj")
+	public String generisiIzvestaj(String idZ) {
+//		TODO:...
+		return "redirect:/guest";
+	}
+	
 	
 
 	
@@ -307,5 +397,10 @@ public class UlogeController {
 		krepo.save(k);
 		
 		return "redirect:/guest";
+	}
+	
+	@GetMapping("/error")
+	public String getErrorPage() {
+		return "accessDenied";
 	}
 }
